Fix issue when /etc/init.d/ doesn't exist in the system

+++ lib/Nipe/Stop.pm
@@ -9,9 +9,13 @@ sub new {
 		system ("sudo iptables -t $table -F OUTPUT");
 		system ("sudo iptables -t $table -F OUTPUT");
 	}
-
-	system("sudo systemctl stop tor");
-
+	
+	if ( -e "/etc/init.d/tor") {
+		system ("sudo /etc/init.d/tor stop > /dev/null");
+	} else { 
+		system ("sudo systemctl stop tor");
+	}
+	
 	return true;
 }
